Almost done.

+++ corewar/src/get_player.c
@@ -6,7 +6,7 @@
 /*   By: tmielcza <tmielcza@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/07/31 11:08:24 by tmielcza          #+#    #+#             */
-/*   Updated: 2015/09/10 16:50:55 by tmielcza         ###   ########.fr       */
+/*   Updated: 2015/09/29 21:42:19 by tmielcza         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ t_player const	*get_player(t_vm const *vm, size_t id)
 	i = 0;
 	while (i < vm->nb_players)
 	{
-		if (vm->players[i].id == id)
+		if (vm->players[i].id == (int)id)
 		{
 			return (vm->players + i);
 		}
